fix type error int to double

+++ API open weather/City.cs
@@ -26,7 +26,7 @@ namespace API_open_weather
         public double temp { get; set; }
         public double feels_like { get; set; }
         public double temp_min { get; set; }
-        public int temp_max { get; set; }
+        public double temp_max { get; set; }
         public int pressure { get; set; }
         public int humidity { get; set; }
     }
